# app.js
import express from 'express'
import logger from 'morgan'
import cors from 'cors'
import swaggerUi from "swagger-ui-express";
import swaggerDocument from './swagger.json'assert { type: 'json' };
import "dotenv/config";

import authRouter from './routes/api/auth.js';
import todayRouter from './routes/api/today.js';
import waterRouter from './routes/api/waterrate.js'

const app = express()

const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short'

app.use(logger(formatsLogger))
app.use(cors())
app.use(function (req, res, next) {
  const allowedOrigins = ['http://localhost:3000', 'https://ihorkoi.github.io/project-wt12-front'];
  const origin = req.headers.origin;
  if (allowedOrigins.includes(origin)) {
    res.setHeader('Access-Control-Allow-Origin', origin);
  }
  res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, Authorization");
  res.header("Access-Control-Allow-credentials", true);
  res.header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, UPDATE");
  next();
});
app.use(express.json())

app.use('/api/auth', authRouter)
app.use('/api/waterrate', waterRouter)
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
app.use('/api/today', todayRouter);

app.use((req, res) => {
  res.status(404).json({ message: 'Not found' })
})

app.use((err, req, res, next) => {
  const { status = 500, message = "Server error" } = err;
  res.status(status).json({ message, })
})

export default app;