# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.1",
    "title": "WaterTraсker API docs",
    "description": "API documentation for project WaterTraсker",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [{ "url": "http://localhost:3000" }],
...
  "paths": {
...
    "/api/waterrate": {
      "post": {
        "tags": ["Water"],
        "summary": "Add water entry",
        "security": [{ "Bearer": [] }],
        "requestBody": {
          "description": "Water entry object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWaterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddWaterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          }
        }
      }
    },
...
    "/api/today": {
      "get": {
        "tags": ["Water"],
        "summary": "Get water entries for today",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddWaterResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddWaterRequest": {
        "type": "object",
        "required": ["waterAmount", "time"],
        "properties": {
          "waterAmount": {
            "type": "string",
            "description": "Amount of water",
            "example": "500"
          },
          "time": {
            "type": "string",
            "description": "Time of water intake",
            "example": "2023-12-01T12:00:00Z"
          }
        }
      },
      "AddWaterResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Water entry ID"
          },
          "waterAmount": {
            "type": "string",
            "description": "Amount of water"
          },
          "time": {
            "type": "string",
            "description": "Time of water intake"
          },
          "owner": {
            "type": "string",
            "description": "ID of the owner user"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}